Graphql example file
