+++ nomos-services/http/examples/axum.rs
@@ -3,9 +3,7 @@ use std::sync::Arc;
 
 use clap::Parser;
 use nomos_http::backends::HttpBackend;
-use nomos_http::bridge::{
-    build_graphql_bridge, build_http_bridge, HttpBridgeRunner, HttpBridgeService,
-};
+use nomos_http::bridge::{build_http_bridge, HttpBridgeRunner, HttpBridgeService};
 use nomos_http::{
     backends::axum::{AxumBackend, AxumBackendSettings},
     http::*,
@@ -104,113 +102,11 @@ where
     }))
 }
 
-fn dummy_graphql_router<B>(
-    handle: overwatch_rs::overwatch::handle::OverwatchHandle,
-) -> HttpBridgeRunner
-where
-    B: HttpBackend + Send + Sync + 'static,
-    B::Error: Error + Send + Sync + 'static,
-{
-    Box::new(Box::pin(async move {
-        let (dummy, mut hello_res_rx) =
-            build_graphql_bridge::<DummyGraphqlService, B, _>(handle, "graphql")
-                .await
-                .unwrap();
-
-        while let Some(GraphqlRequest { res_tx, req }) = hello_res_rx.recv().await {
-            let (sender, receiver) = oneshot::channel();
-            dummy
-                .send(DummyGraphqlMsg {
-                    req,
-                    reply_channel: sender,
-                })
-                .await
-                .unwrap();
-            let res = receiver.await.unwrap();
-            res_tx.send(res).await.unwrap();
-        }
-        Ok(())
-    }))
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct DummyGraphql {
-    val: Arc<Mutex<i32>>,
-}
-
-#[async_graphql::Object]
-impl DummyGraphql {
-    async fn val(&self) -> i32 {
-        let mut val = self.val.lock();
-        *val += 1;
-        *val
-    }
-}
-
-pub struct DummyGraphqlService {
-    service_state: ServiceStateHandle<Self>,
-}
-
-#[derive(Debug)]
-pub struct DummyGraphqlMsg {
-    req: async_graphql::Request,
-    reply_channel: oneshot::Sender<async_graphql::Response>,
-}
-
-impl RelayMessage for DummyGraphqlMsg {}
-
-impl ServiceData for DummyGraphqlService {
-    const SERVICE_ID: ServiceId = "DummyGraphqlService";
-    type Settings = ();
-    type State = NoState<()>;
-    type StateOperator = NoOperator<Self::State>;
-    type Message = DummyGraphqlMsg;
-}
-
-#[async_trait::async_trait]
-impl ServiceCore for DummyGraphqlService {
-    fn init(service_state: ServiceStateHandle<Self>) -> Result<Self, overwatch_rs::DynError> {
-        Ok(Self { service_state })
-    }
-
-    async fn run(self) -> Result<(), overwatch_rs::DynError> {
-        let Self {
-            service_state: ServiceStateHandle {
-                mut inbound_relay, ..
-            },
-        } = self;
-
-        static SCHEMA: once_cell::sync::Lazy<
-            async_graphql::Schema<
-                DummyGraphql,
-                async_graphql::EmptyMutation,
-                async_graphql::EmptySubscription,
-            >,
-        > = once_cell::sync::Lazy::new(|| {
-            async_graphql::Schema::build(
-                DummyGraphql::default(),
-                async_graphql::EmptyMutation,
-                async_graphql::EmptySubscription,
-            )
-            .finish()
-        });
-
-        // Handle the http request to dummy service.
-        while let Some(msg) = inbound_relay.recv().await {
-            let res = SCHEMA.execute(msg.req).await;
-            msg.reply_channel.send(res).unwrap();
-        }
-
-        Ok(())
-    }
-}
-
 #[derive(overwatch_derive::Services)]
 struct Services {
     http: ServiceHandle<HttpService<AxumBackend>>,
     router: ServiceHandle<HttpBridgeService>,
     dummy: ServiceHandle<DummyService>,
-    dummy_graphql: ServiceHandle<DummyGraphqlService>,
 }
 
 #[derive(clap::Parser)]
@@ -229,13 +125,9 @@ fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
                 backend: settings.http,
             },
             router: nomos_http::bridge::HttpBridgeSettings {
-                runners: vec![
-                    Arc::new(Box::new(dummy_router::<AxumBackend>)),
-                    Arc::new(Box::new(dummy_graphql_router::<AxumBackend>)),
-                ],
+                runners: vec![Arc::new(Box::new(dummy_router::<AxumBackend>))],
             },
             dummy: (),
-            dummy_graphql: (),
         },
         None,
     )?;
